requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
django-phonenumber-field[phonenumbers]
django-jalali
django-jalali[drf]
django-admin-interface
django-filter
drf-yasg
drf-spectacular
jdatetime
pytest-django
psycopg2-binary
isort
ruff
python-decouple
Faker
Pillow
